requirements.txt
# Core Flask dependencies
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1

# Database
SQLAlchemy==2.0.41
psycopg2-binary==2.9.9

# Security
Werkzeug==3.1.3
itsdangerous==2.2.0

# HTTP requests
requests==2.32.4

# Templating
Jinja2==3.1.6
MarkupSafe==3.0.2

# Image processing
Pillow==10.1.0

# Environment and production
python-dotenv==1.0.0
gunicorn==21.2.0

# Security and rate limiting
flask-limiter==3.5.0
email-validator==2.1.0
